Fix dupe bugs in the SAG Mill with grinding balls.
This will fix all of the dupe bugs in the current version with grinding items with grinding balls and getting extras. E.g. blocks into dusts and getting extra. (The same issue as #388 it just applies to more things than quartz, quartz issue is also a problem due to being able to turn dusts into pure crystals and back into normal.)

The stonebrick into cobblestone will also cause the same behaviour however I didn't include them in this since their recipes already allow for duplication as it is. (It's also just cobblestone, easier ways of getting mass amounts of it.)