Fix IC2 Coolant Cell Recipe.
Coolant cells are expected to have a damage value of 1, all other recipes using them are only accepting them with a damage value of 1.
This fixes the recipe to return them with that damage value directly, instead of having to place the undamaged one in a crafting grid, to get one with damage value 1.